requirements.txt
streamlit
pandas
numpy
scipy
tensorflow-cpu
scikit-learn == 1.0.2